Restructuring

+++ contracts/IVBEP20.sol
@@ -9,7 +9,7 @@ interface IVBEP20 {
 
     function supplyRatePerBlock() external returns (uint256);
 
-    function redeem(uint) external returns (uint);
+    function redeem(uint256) external returns (uint256);
 
-    function redeemUnderlying(uint) external returns (uint);
+    function redeemUnderlying(uint256) external returns (uint256);
 }
